move eqloquent models
