+++ infra/EloquentModels/AdminUser.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Infra\EloquentModels;
+
+use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Notifications\Notifiable;
+
+class AdminUser extends Authenticatable
+{
+    use Notifiable;
+
+    /**
+     * The table associated with the model
+     *
+     * @var string
+     */
+    protected $table = 'admin_users';
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $guarded = [
+        'id',
+    ];
+
+    /**
+     * The attributes that should be hidden for arrays.
+     *
+     * @var array
+     */
+    protected $hidden = [
+        'password',
+    ];
+
+    protected $casts = [
+        'status' => 'boolean',
+    ];
+}
